fix(swc-angular): 🐞 ignore .swcrc to fix conflicts
without disabling .swcrc parsing, users will likely stumble upon
the following error when an .swcrc file is present:
`env` and `jsc.target` cannot be used together
